\name{Mrklist}
\docType{data}
\alias{Mrklist}
\title{A list of testing marker}
\description{A tab-delimited file that lists all SNP markers included in the
 CCdata file without a header line. Each row in this file represents one marker 
and includes three fields: marker name, chromosome number, map position (physical or genetic)
}
\usage{Mrklist}
\format{A tab-delimited file without header.}
\keyword{Mrklist}