\name{OSACC-package}
\alias{OSACC-package}
\alias{OSACC, OSACC1, OSACC2}
\docType{package}
\title{
Ordered Subset Analysis for Case-Control Studies
}
\description{
The R package consists of three functions, OSACC, OSACC1 and OSACC2.  OSACC is  C ++ code implemented in R. OSACC1 and OSACC2 are written in R, which makes it easy to modify the source code when the program user wishes to use different test statistics or to incorporate additional covariates in a logistic regression model. OSACC has options for all four tests, OSACC1-risk, OSACC1-hom, OSACC2-risk, OSACC2-hom. OSACC1 has options for two tests, OSACC1-risk, OSACC1-hom. OSACC2 has options for two tests, OSACC2-risk, OSACC2-hom. Please consult the reference for a detailed description of the respective null hypotheses being tested. 
}
\details{
\tabular{ll}{
Package: \tab OSACC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-12-11\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{ Xuejun Qin, Elizabeth R. Hauser, Silke Schmidt, Center for Human Genetics, Duke University

Maintainer: Xuejun Qin<xjq@chg.duhs.duke.edu>
}
\references{Xuejun Qin, Elizabeth R. Hauser, Silke Schmidt (2010), Ordered Subset Analysis for Case-Control Studies, Genetic Epidemiology 34(5), 407-417.. 
}
\keyword{OSACC, OSACC1, OSACC2}

